+++ src/aiohttp_things/views.py
@@ -3,6 +3,7 @@ Reserved for simple views for traversal routing.
 
 More complex classes must naming web handlers.
 """
+
 from aiohttp_things.web_handlers import (
     ContextMixin,
     ItemMixin,
@@ -15,12 +16,12 @@ from aiohttp_things.web_handlers import (
 )
 
 __all__ = [
-    'ContextMixin',
-    'ItemMixin',
-    'Jinja2Mixin',
-    'JSONMixin',
-    'ListMixin',
-    'PaginationMixin',
-    'PrimaryKeyMixin',
-    'ResponseFormatMixin',
+    "ContextMixin",
+    "ItemMixin",
+    "Jinja2Mixin",
+    "JSONMixin",
+    "ListMixin",
+    "PaginationMixin",
+    "PrimaryKeyMixin",
+    "ResponseFormatMixin",
 ]